% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicBoxPlot.R
\name{classicBoxPlot}
\alias{classicBoxPlot}
\title{Computes classical box-plots.}
\usage{
classicBoxPlot(x, ylab = NULL)
}
\arguments{
\item{x}{A vector. The data set under study.}

\item{ylab}{A character. The label to be printed in the y axis of eventual
plots.}
}
\value{
characteristics of the boxplot
}
\description{
Is based on quartiles.
}
\examples{
classicBoxPlot(rnorm(100), ylab="example")
}
