% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribFreq.R
\name{distFreq}
\alias{distFreq}
\title{Para construção de distribuição de frequências nos moldes da disciplina.}
\usage{
distFreq(x, varName = "")
}
\arguments{
\item{x}{Um conjunto de dados, em forma de vetor. Os dados sobre os quais deseja-se a distribuição de frequências.}

\item{varName}{Um texto, representando o nome da variável sob estudo.}
}
\value{
Uma lista, contemplando:

qualitativeData - Os dados qualitativizados

distFreq - A distribuição de freuências

grid -  A grade de pontos que compõem as categorias, para o caso de quantitaitvas
}
\description{
Dá tratamento especial para cada tipo de variável e recorre ao método mais simples para quantitativas.
}
\examples{
distFreq(x = c(x = runif(100), varName = "exemplo"))
}
