% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_function.R
\name{ourMode}
\alias{ourMode}
\title{A simple function for obtaining a single mode of a given sample.}
\usage{
ourMode(
  sample,
  xlab = "",
  toPrint = TRUE,
  minimalAmplitudeRatioForGrouping = 0.1
)
}
\arguments{
\item{sample}{A vector. The data set under study.}

\item{xlab}{A character. The label to be printed in the x axis of eventual
plots.}

\item{toPrint}{A logical. If TRUE, then plots and frequency distributions are
exhibited.}

\item{minimalAmplitudeRatioForGrouping}{A numeric. For quantitative
variables, if \code{amplitude(sample)/length(sample)} is lesser than
minimalAmplitudeRatioForGrouping) then no subinterval construction of the
data is necessary.}
}
\value{
A vector, numeric or character, depending on the nature of
  \code{sample}. The respective mode.
}
\description{
For quantitative variables, the Czuber's mode can be adopted, depending on
the value of the ratio: total amplitude divided by the sample size. It is
supposed that just one category/subinterval is the most frequent. For general
purposes, sophistication must be introduced in the code. the R package
\code{\link{modeest}} also involves a number of functions for mode
calculation.
}
\examples{
x = rnorm(100)
ourMode(sample = x, xlab = "example", toPrint = TRUE)
}
